Network Scan Folder does not work [#318 responsible:none state:fixed]
Updated so that SOLA can connect to a network share using NTLM credentails.
Also enhanced the feature so that filters can be used to only display
specific documents as well as allowing files from all subfolders of the
Network Share that match the filters to be displayed.

Also updated the SharedFolderCleaner EJB to use the new file share functionality.