Password Encryption Issues [#410 responsible:none state:fixed]
SOLA uses the stock standard MessageDigest algorithm to compute the password hash and then coverts the hash to a hexidecimal string using the BigInteger class. In some cases (approx 6% of passwords) this approach truncates a leading 0 from the Hex string making it invalid. Updated to use a different hex conversion method.