Updates to mail service [#418]
Modified email service configuration so that it does not
fail deployment of the SOLA web services if the developer
has not configured the mail/sola JNDI resource.
Also modified so that the configuration parameters are
retrieved each time the service is executed so that any
changed configuration values are picked up without needing
to restart the SOLA Web Services.