Further updates to Redact functionality [#3 responsible:none state:fixed]
Updated so that child entities as well as child entity lists can be
marked for redaction. Redacting a child entity (or list) from a parent
is the same as applying the security classification directly to the
child (or each child in the child list). The child entity will not be
loaded so a child field will be null and child list fields will be empty.